Fix: GET SNMP no longer returns shard
Many of the scripts in this repository used the GET SNMP settings endpoint to resolve the Meraki dashboard shard where the organization to be interacted with is stored and make calls to the shard directly and not api.meraki.com. The intent of this technique was to gain a tiny bit of performance. However, for most organizations the GET SNMP settings endpoint returns "snmp.meraki.com" nowadays, which does not respont to API calls. To work around this change, the scripts have been patched to use the new API mega proxy instead (api-mp.meraki.com), which should give them similar performance, while operating in a more standard way. If a script you are using has stopped working due to this change, please open an issue about it.